move ViewModel code
